+++ huaweicloud-sdk-java-dis-iface/src/main/java/com/huaweicloud/dis/iface/data/request/DeleteCheckpointRequest.java
@@ -1,19 +1,3 @@
-/*
- * Copyright 2002-2010 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package com.huaweicloud.dis.iface.data.request;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
@@ -22,8 +6,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public class GetCheckpointRequest
-{
+public class DeleteCheckpointRequest {
+
     @JsonProperty("stream_name")
     private String streamName;
 
@@ -35,9 +19,6 @@ public class GetCheckpointRequest
 
     @JsonProperty("app_name")
     private String appName;
-
-    @JsonProperty("timestamp")
-    private Long timeStamp;
 
     public String getStreamName()
     {
@@ -79,20 +60,11 @@ public class GetCheckpointRequest
         this.appName = appId;
     }
 
-    public Long getTimeStamp()
-    {
-        return timeStamp;
-    }
-
-    public void setTimeStamp(Long timeStamp)
-    {
-        this.timeStamp = timeStamp;
-    }
 
     @Override
     public String toString()
     {
-        return "GetCheckpointParam [streamName=" + streamName + ", shardId=" + partitionId + ", checkpointType="
-                + checkpointType + ", appId=" + appName  + ", timeStamp=" + timeStamp +  "]";
+        return "DeleteCheckpointParam [streamName=" + streamName + ", shardId=" + partitionId + ", checkpointType="
+                + checkpointType + ", appId=" + appName  +  "]";
     }
 }