+++ huaweicloud-sdk-java-dis-iface/src/main/java/com/huaweicloud/dis/iface/data/response/DeleteCheckpointResult.java
@@ -1,0 +1,9 @@
+package com.huaweicloud.dis.iface.data.response;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class DeleteCheckpointResult {
+}