+++ huaweicloud-sdk-java-dis/src/main/java/com/huaweicloud/dis/DISClient.java
@@ -22,6 +22,10 @@ import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.locks.ReentrantLock;
 
+import com.huaweicloud.dis.iface.app.request.ListStreamConsumingStateRequest;
+import com.huaweicloud.dis.iface.app.response.ListStreamConsumingStateResult;
+import com.huaweicloud.dis.iface.data.request.*;
+import com.huaweicloud.dis.iface.data.response.*;
 import org.apache.http.HttpRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -45,21 +49,6 @@ import com.huaweicloud.dis.iface.app.request.CreateAppRequest;
 import com.huaweicloud.dis.iface.app.request.ListAppsRequest;
 import com.huaweicloud.dis.iface.app.response.DescribeAppResult;
 import com.huaweicloud.dis.iface.app.response.ListAppsResult;
-import com.huaweicloud.dis.iface.data.request.CommitCheckpointRequest;
-import com.huaweicloud.dis.iface.data.request.GetCheckpointRequest;
-import com.huaweicloud.dis.iface.data.request.GetPartitionCursorRequest;
-import com.huaweicloud.dis.iface.data.request.GetRecordsRequest;
-import com.huaweicloud.dis.iface.data.request.PutRecordRequest;
-import com.huaweicloud.dis.iface.data.request.PutRecordsRequest;
-import com.huaweicloud.dis.iface.data.request.QueryFileState;
-import com.huaweicloud.dis.iface.data.response.CommitCheckpointResult;
-import com.huaweicloud.dis.iface.data.response.FileUploadResult;
-import com.huaweicloud.dis.iface.data.response.GetCheckpointResult;
-import com.huaweicloud.dis.iface.data.response.GetPartitionCursorResult;
-import com.huaweicloud.dis.iface.data.response.GetRecordsResult;
-import com.huaweicloud.dis.iface.data.response.PutRecordResult;
-import com.huaweicloud.dis.iface.data.response.PutRecordsResult;
-import com.huaweicloud.dis.iface.data.response.PutRecordsResultEntry;
 import com.huaweicloud.dis.iface.stream.request.CreateStreamRequest;
 import com.huaweicloud.dis.iface.stream.request.DeleteStreamRequest;
 import com.huaweicloud.dis.iface.stream.request.DescribeStreamRequest;
@@ -596,5 +585,44 @@ public class DISClient extends AbstractDISClient implements DIS
         setEndpoint(request, disConfig.getEndpoint());
         return request(getCheckpointRequest, request, GetCheckpointResult.class);
     }
-    
+
+
+    public DeleteCheckpointResult deleteCheckpoint(DeleteCheckpointRequest deleteCheckpointRequest)
+    {
+        return innerDeleteCheckpoint(deleteCheckpointRequest);
+    }
+
+    protected final DeleteCheckpointResult innerDeleteCheckpoint(DeleteCheckpointRequest deleteCheckpointRequest)
+    {
+        Request<HttpRequest> request = new DefaultRequest<>(Constants.SERVICENAME);
+        request.setHttpMethod(HttpMethodName.DELETE);
+
+        final String resourcePath = ResourcePathBuilder.standard()
+                .withProjectId(disConfig.getProjectId())
+                .withResource(new CheckPointResource(null))
+                .build();
+        request.setResourcePath(resourcePath);
+        setEndpoint(request, disConfig.getEndpoint());
+        return request(deleteCheckpointRequest, request, DeleteCheckpointResult.class);
+    }
+
+    @Override
+    public ListStreamConsumingStateResult listStreamConsumingState(ListStreamConsumingStateRequest listStreamConsumingStateRequest) {
+        return innerListStreamConsumingState(listStreamConsumingStateRequest);
+    }
+
+    protected ListStreamConsumingStateResult innerListStreamConsumingState(ListStreamConsumingStateRequest listStreamConsumingStateRequest) {
+        Request<HttpRequest> request = new DefaultRequest<>(Constants.SERVICENAME);
+        request.setHttpMethod(HttpMethodName.GET);
+
+        final String resourcePath = ResourcePathBuilder.standard()
+                .withProjectId(disConfig.getProjectId())
+                .withResource(new AppsResource(listStreamConsumingStateRequest.getAppName()))
+                .withResource(new StreamResource(listStreamConsumingStateRequest.getStreamName()))
+                .build();
+        request.setResourcePath(resourcePath);
+        setEndpoint(request, disConfig.getManagerEndpoint());
+        return request(null, request, ListStreamConsumingStateResult.class);
+    }
+
 }