+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 - Features
   * add describeApp and listApps features
   * use NIO for DISClientAsync 
+  * add delete checkpoint
+  * add describe consumer state